+++ examples/factory_async.rs
@@ -88,9 +88,9 @@ impl AsyncFactoryComponent for Counter {
         }
     }
 
-    fn init_loading_widgets(root: &mut Self::Root) -> Option<LoadingWidgets> {
+    fn init_loading_widgets(root: Self::Root) -> Option<LoadingWidgets> {
         view! {
-            #[local_ref]
+            #[local]
             root {
                 set_orientation: gtk::Orientation::Horizontal,
                 set_spacing: 10,