+++ examples/libadwaita/tab_factory.rs
@@ -157,7 +157,7 @@ impl SimpleComponent for App {
 
     fn init(
         counter: Self::Init,
-        root: &Self::Root,
+        root: Self::Root,
         sender: ComponentSender<Self>,
     ) -> ComponentParts<Self> {
         let counters = FactoryVecDeque::builder()