+++ relm4-components/src/web_image.rs
@@ -41,7 +41,7 @@ impl Component for WebImage {
     ) -> ComponentParts<Self> {
         let widget = gtk::Box::default();
         root.append(&widget);
-        let current_widget = Self::set_spinner(root, widget.upcast_ref());
+        let current_widget = Self::set_spinner(&root, widget.upcast_ref());
 
         let model = Self {
             current_id: 0,